refactor(backend): move user prefs to dedicated module
