+++ apps/backend/src/models.rs
@@ -3,62 +3,6 @@ use chrono::NaiveDate;
 use sea_orm::FromJsonQueryResult;
 use serde::{Deserialize, Serialize};
 
-
-
-#[derive(
-    Debug,
-    Serialize,
-    Deserialize,
-    SimpleObject,
-    Clone,
-    InputObject,
-    Eq,
-    PartialEq,
-    FromJsonQueryResult,
-)]
-pub struct UserFeaturesEnabledPreferences {
-    pub anime: bool,
-    pub audio_books: bool,
-    pub books: bool,
-    pub manga: bool,
-    pub movies: bool,
-    pub podcasts: bool,
-    pub shows: bool,
-    pub video_games: bool,
-}
-
-impl Default for UserFeaturesEnabledPreferences {
-    fn default() -> Self {
-        Self {
-            anime: true,
-            audio_books: true,
-            books: true,
-            manga: true,
-            movies: true,
-            podcasts: true,
-            shows: true,
-            video_games: true,
-        }
-    }
-}
-
-#[derive(
-    Debug,
-    Serialize,
-    Deserialize,
-    SimpleObject,
-    Clone,
-    InputObject,
-    Eq,
-    PartialEq,
-    Default,
-    FromJsonQueryResult,
-)]
-pub struct UserPreferences {
-    #[serde(default)]
-    pub features_enabled: UserFeaturesEnabledPreferences,
-}
-
 pub mod media {
     use super::*;
